Add testing & refactor folder structure
